+++ app/src/main/java/protect/card_locker/AboutActivity.java
@@ -1,11 +1,13 @@
 package protect.card_locker;
 
+import android.content.Intent;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
+import android.net.Uri;
 import android.os.Bundle;
-import android.text.method.LinkMovementMethod;
 import android.util.Log;
 import android.view.MenuItem;
+import android.view.View;
 import android.widget.TextView;
 
 import java.io.BufferedReader;
@@ -17,12 +19,15 @@ import java.util.Calendar;
 import java.util.List;
 
 import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.widget.Toolbar;
+import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.core.text.HtmlCompat;
 
-public class AboutActivity extends CatimaAppCompatActivity
+public class AboutActivity extends CatimaAppCompatActivity implements View.OnClickListener
 {
     private static final String TAG = "Catima";
+    ConstraintLayout version_history, translate, license, repo, privacy, error, credits, rate;
 
     @Override
     protected void onCreate(Bundle savedInstanceState)
@@ -93,39 +98,82 @@ public class AboutActivity extends CatimaAppCompatActivity
             Log.w(TAG, "Package name not found", e);
         }
 
+        TextView copyright = findViewById(R.id.credits_sub);
+        copyright.setText(String.format(getString(R.string.app_copyright_fmt), year));
+        TextView vHistory = findViewById(R.id.version_history_sub);
+        vHistory.setText(String.format(getString(R.string.debug_version_fmt), version));
+
         setTitle(String.format(getString(R.string.about_title_fmt), appName));
 
-        TextView aboutTextView = findViewById(R.id.aboutText);
-        aboutTextView.setText(HtmlCompat.fromHtml(String.format(getString(R.string.debug_version_fmt), version) +
-                "<br/><br/>" +
-                String.format(getString(R.string.app_revision_fmt),
-                        "<a href=\"" + getString(R.string.app_revision_url) + "\">" +
-                                "GitHub" +
-                                "</a>") +
-                "<br/><br/>" +
-                String.format(getString(R.string.app_copyright_fmt), year) +
-                "<br/><br/>" +
-                getString(R.string.app_copyright_old) +
-                "<br/><br/>" +
-                getString(R.string.app_license) +
-                "<br/><br/>" +
-                String.format(getString(R.string.app_contributors), contributors.toString()) +
-                "<br/><br/>" +
-                String.format(getString(R.string.app_libraries), libs.toString()) +
-                "<br/><br/>" +
-                String.format(getString(R.string.app_resources), resources.toString()), HtmlCompat.FROM_HTML_MODE_COMPACT));
-        aboutTextView.setMovementMethod(LinkMovementMethod.getInstance());
+        version_history = findViewById(R.id.version_history);
+        translate = findViewById(R.id.translate);
+        license = findViewById(R.id.license);
+        repo = findViewById(R.id.repo);
+        privacy = findViewById(R.id.privacy);
+        error = findViewById(R.id.report_error);
+        credits = findViewById(R.id.credits);
+        rate = findViewById(R.id.rate);
+
+        version_history.setOnClickListener(this);
+        translate.setOnClickListener(this);
+        license.setOnClickListener(this);
+        repo.setOnClickListener(this);
+        privacy.setOnClickListener(this);
+        error.setOnClickListener(this);
+        rate.setOnClickListener(this);
+
+        StringBuilder contributorInfo = new StringBuilder();
+        contributorInfo.append(HtmlCompat.fromHtml(String.format(getString(R.string.app_contributors), contributors.toString()), HtmlCompat.FROM_HTML_MODE_COMPACT));
+        contributorInfo.append("\n\n");
+        contributorInfo.append(getString(R.string.app_copyright_old));
+        contributorInfo.append("\n\n");
+        contributorInfo.append(HtmlCompat.fromHtml(String.format(getString(R.string.app_libraries), libs.toString()), HtmlCompat.FROM_HTML_MODE_COMPACT));
+        contributorInfo.append("\n\n");
+        contributorInfo.append(HtmlCompat.fromHtml(String.format(getString(R.string.app_resources), resources.toString()), HtmlCompat.FROM_HTML_MODE_COMPACT));
+
+        credits.setOnClickListener(view -> new AlertDialog.Builder(this)
+                .setTitle(R.string.credits)
+                .setMessage(contributorInfo.toString())
+                .setPositiveButton(R.string.ok, (dialogInterface, i) -> {})
+                .show());
     }
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item)
     {
         int id = item.getItemId();
-
         if (id == android.R.id.home) {
             finish();
         }
-
         return super.onOptionsItemSelected(item);
     }
+
+    @Override
+    public void onClick(View view) {
+        int id = view.getId();
+
+        String url;
+        if (id == R.id.version_history) {
+            url = "https://catima.app/changelog/";
+        } else if (id == R.id.translate) {
+            url = "https://hosted.weblate.org/engage/catima/";
+        } else if (id == R.id.license) {
+            url = "https://github.com/TheLastProject/Catima/blob/master/LICENSE";
+        } else if (id == R.id.repo) {
+            url = "https://github.com/TheLastProject/Catima/";
+        } else if (id == R.id.privacy) {
+            url = "https://catima.app/privacy-policy/";
+        } else if (id == R.id.report_error) {
+            url = "https://github.com/TheLastProject/Catima/issues";
+        } else if (id == R.id.rate) {
+            url = "https://play.google.com/store/apps/details?id=me.hackerchick.catima";
+        } else {
+            return;
+        }
+
+        Intent intent = new Intent(Intent.ACTION_VIEW);
+        intent.setData(Uri.parse(url));
+        startActivity(intent);
+    }
+
 }