+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased - 85
+
+Android 4.4 is no longer supported starting with this release. If you want to use Catima on Android 4.4, please use version 2.6.1.
+
+- Improved Android 12 support
+- Improved about screen
+- Search now ignores accents
+
 ## v2.6.1 - 84 (2021-09-25)
 
 - Minor bugfixes and improvements